backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.3
django-filter==23.2
django-import-export==3.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.3
numpy==1.25.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytils==0.4.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
psycopg2-binary==2.9.3